refactor: centralize audio modules
